Move redisClient, remove vscode config, fix types
