+++ .vscode/launch.json
@@ -28,6 +28,7 @@
         "--inspect-brk",
         "${workspaceRoot}/node_modules/.bin/jest",
         "--runInBand",
+        "--coverage",
         "false"
       ],
       "console": "integratedTerminal"
@@ -43,23 +44,7 @@
         "--inspect-brk",
         "${workspaceRoot}/node_modules/.bin/jest",
         "--runInBand",
-        "false",
-        "${relativeFile}"
-      ],
-      "console": "integratedTerminal"
-  },
-  {
-      "type": "node",
-      "request": "launch",
-      "name": "Debug current E2E test file",
-      "skipFiles": [
-          "<node_internals>/**"
-      ],
-      "runtimeArgs": [
-        "--inspect-brk",
-        "${workspaceRoot}/node_modules/.bin/jest",
-        "--runInBand",
-        "--config=test/jest-ci.json",
+        "--coverage",
         "false",
         "${relativeFile}"
       ],
