+++ src/routes/safe-apps/__tests__/get-safe-apps.e2e-spec.ts
@@ -1,10 +1,11 @@
 import * as request from 'supertest';
+import { RedisClientType } from 'redis';
 import { INestApplication } from '@nestjs/common';
 import { Test } from '@nestjs/testing';
 import { AppModule } from '../../../app.module';
 import { readFileSync } from 'fs';
-import { RedisClientType, createClient } from 'redis';
 import { SafeApp } from '../entities/safe-app.entity';
+import { redisClientFactory } from '../../../../test/common/cache/redis-client.factory';
 
 describe('Get Safe Apps e2e test', () => {
   let app: INestApplication;
@@ -83,12 +84,3 @@ describe('Get Safe Apps e2e test', () => {
     await redisClient.quit();
   });
 });
-
-async function redisClientFactory(): Promise<RedisClientType> {
-  const { REDIS_HOST = 'localhost', REDIS_PORT = 6379 } = process.env;
-  const client: RedisClientType = createClient({
-    url: `redis://${REDIS_HOST}:${REDIS_PORT}`,
-  });
-  client.connect();
-  return client;
-}
