+++ src/routes/safe-apps/entities/safe-app.entity.ts
@@ -1,4 +1,4 @@
-import { ApiProperty } from '@nestjs/swagger';
+import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
 import { SafeApp as DomainSafeApp } from '../../../domain/safe-apps/entities/safe-app.entity';
 import { SafeAppAccessControl } from './safe-app-access-control.entity';
 import { SafeAppProvider } from './safe-app-provider.entity';
@@ -16,7 +16,7 @@ export class SafeApp implements DomainSafeApp {
   description: string;
   @ApiProperty()
   chainIds: string[];
-  @ApiProperty()
+  @ApiPropertyOptional()
   provider?: SafeAppProvider;
   @ApiProperty()
   accessControl: SafeAppAccessControl;
